Example app controls
- Add vertical view pager
- Add controls fragment
- Separate parent and children gridlayout from controls
- Clean up continuous weighted weights
