--- app/src/main/res/layout/activity_fragment.xml
@@ -1,174 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~     Spruce
-  ~
-  ~     Copyright (c) 2017 WillowTree, Inc.
-  ~     Permission is hereby granted, free of charge, to any person obtaining a copy
-  ~     of this software and associated documentation files (the "Software"), to deal
-  ~     in the Software without restriction, including without limitation the rights
-  ~     to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-  ~     copies of the Software, and to permit persons to whom the Software is
-  ~     furnished to do so, subject to the following conditions:
-  ~     The above copyright notice and this permission notice shall be included in
-  ~     all copies or substantial portions of the Software.
-  ~     THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-  ~     IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-  ~     FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-  ~     AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-  ~     LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-  ~     OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-  ~     THE SOFTWARE.
-  ~
-  -->
-
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              android:id="@+id/view_fragment"
-              android:layout_width="match_parent"
-              android:layout_height="match_parent"
-              android:orientation="vertical">
-
-    <android.support.v7.widget.Toolbar
-        android:id="@+id/tool_bar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:theme="@style/LightToolbarTheme"
-        android:background="@color/colorPrimary">
-
-        <Spinner
-            android:id="@+id/sort_selection"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:background="@null"
-            android:gravity="start"/>
-
-    </android.support.v7.widget.Toolbar>
-
-    <FrameLayout
-        android:layout_width="wrap_content"
-        android:layout_height="0dp"
-        android:layout_gravity="center"
-        android:layout_weight="2">
-
-        <GridLayout
-            android:id="@+id/view_group_to_animate"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_gravity="center"
-            android:layout_marginBottom="4dp"
-            android:layout_marginLeft="4dp"
-            android:layout_marginRight="4dp"
-            android:layout_marginTop="4dp"
-            android:clickable="true"
-            android:columnCount="11"
-            android:rowCount="11"/>
-
-    </FrameLayout>
-
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="1dp"
-        android:background="@color/lighter_gray"/>
-
-    <ScrollView
-        android:id="@+id/control_view"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_marginBottom="4dp"
-        android:layout_marginLeft="4dp"
-        android:layout_marginRight="4dp"
-        android:layout_weight=".5">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical">
-
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:orientation="horizontal">
-
-                <include layout="@layout/directional_radio_group" />
-
-                <include layout="@layout/positional_radio_group" />
-
-                <include layout="@layout/continuous_radio_group" />
-
-                <include layout="@layout/cornered_radio_group" />
-
-            </LinearLayout>
-
-            <CheckBox
-                android:id="@+id/linear_reversed"
-                android:layout_width="wrap_content"
-                android:layout_height="0dp"
-                android:layout_weight="1"
-                android:text="@string/reversed"
-                android:visibility="gone"/>
-
-            <TextView
-                android:id="@+id/seek_bar_title"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginLeft="4dp"
-                android:layout_marginBottom="4dp"
-                android:text="@string/duration"/>
-
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:orientation="horizontal"
-                android:weightSum="10"
-                android:layout_gravity="center">
-
-                <TextView
-                    android:id="@+id/animation_start"
-                    android:layout_width="0dp"
-                    android:layout_height="wrap_content"
-                    android:layout_marginLeft="8dp"
-                    android:layout_marginRight="4dp"
-                    android:layout_weight="1"
-                    android:gravity="start"
-                    android:text="@string/animation_start_duration"/>
-
-                <SeekBar
-                    android:id="@+id/animation_seek"
-                    android:layout_width="0dp"
-                    android:layout_height="wrap_content"
-                    android:layout_weight="8"
-                    android:progress="25"/>
-
-                <TextView
-                    android:id="@+id/animation_end"
-                    android:layout_width="0dp"
-                    android:layout_height="wrap_content"
-                    android:layout_marginLeft="4dp"
-                    android:layout_marginRight="8dp"
-                    android:layout_weight="1"
-                    android:gravity="end"
-                    android:text="@string/animation_end_duration"/>
-
-            </LinearLayout>
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="4dp"
-                android:layout_marginBottom="4dp"
-                android:text="@string/code_sample"/>
-
-            <EditText
-                android:id="@+id/code_sample"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="4dp"
-                android:layout_marginBottom="4dp"
-                android:inputType="none"
-                android:textIsSelectable="true"
-                android:background="@drawable/code_sample_background"/>
-
-        </LinearLayout>
-
-    </ScrollView>
-
-</LinearLayout>
