+++ app/src/main/res/layout/continuous_radio_group.xml
@@ -22,88 +22,95 @@
   -->
 
 <merge xmlns:android="http://schemas.android.com/apk/res/android"
-       android:layout_width="wrap_content"
+       android:layout_width="match_parent"
        android:layout_height="wrap_content">
 
     <LinearLayout
-        android:id="@+id/vertical_weight"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:visibility="gone"
-        android:orientation="vertical">
+        android:orientation="horizontal">
 
-        <TextView
+        <LinearLayout
+            android:id="@+id/vertical_weight"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:text="@string/vertical"/>
+            android:visibility="gone"
+            android:orientation="vertical">
 
-        <RadioGroup
-            android:id="@+id/vertical_weighted_radio_group"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content">
-
-            <RadioButton
-                android:id="@+id/vertical_light"
+            <TextView
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:text="@string/light"/>
+                android:layout_gravity="center"
+                android:text="@string/vertical"/>
 
-            <RadioButton
-                android:id="@+id/vertical_medium"
+            <RadioGroup
+                android:id="@+id/vertical_weighted_radio_group"
                 android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:checked="true"
-                android:text="@string/medium"/>
+                android:layout_height="wrap_content">
 
-            <RadioButton
-                android:id="@+id/vertical_heavy"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="@string/heavy"/>
+                <RadioButton
+                    android:id="@+id/vertical_light"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:text="@string/light"/>
 
-        </RadioGroup>
+                <RadioButton
+                    android:id="@+id/vertical_medium"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:checked="true"
+                    android:text="@string/medium"/>
 
-    </LinearLayout>
+                <RadioButton
+                    android:id="@+id/vertical_heavy"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:text="@string/heavy"/>
 
-    <LinearLayout
-        android:id="@+id/horizontal_weight"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:visibility="gone"
-        android:orientation="vertical">
+            </RadioGroup>
 
-        <TextView
+        </LinearLayout>
+
+        <LinearLayout
+            android:id="@+id/horizontal_weight"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:text="@string/horizontal"/>
+            android:visibility="gone"
+            android:orientation="vertical">
 
-        <RadioGroup
-            android:id="@+id/horizontal_weighted_radio_group"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content">
-
-            <RadioButton
-                android:id="@+id/horizontal_light"
+            <TextView
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:text="@string/light"/>
+                android:layout_gravity="center"
+                android:text="@string/horizontal"/>
 
-            <RadioButton
-                android:id="@+id/horizontal_medium"
+            <RadioGroup
+                android:id="@+id/horizontal_weighted_radio_group"
                 android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:checked="true"
-                android:text="@string/medium"/>
+                android:layout_height="wrap_content">
 
-            <RadioButton
-                android:id="@+id/horizontal_heavy"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="@string/heavy"/>
+                <RadioButton
+                    android:id="@+id/horizontal_light"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:text="@string/light"/>
 
-        </RadioGroup>
+                <RadioButton
+                    android:id="@+id/horizontal_medium"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:checked="true"
+                    android:text="@string/medium"/>
+
+                <RadioButton
+                    android:id="@+id/horizontal_heavy"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:text="@string/heavy"/>
+
+            </RadioGroup>
+
+        </LinearLayout>
 
     </LinearLayout>
