+++ app/src/main/res/layout/positional_radio_group.xml
@@ -22,8 +22,8 @@
   -->
 
 <merge xmlns:android="http://schemas.android.com/apk/res/android"
-       android:layout_width="match_parent"
-       android:layout_height="match_parent">
+       android:layout_width="wrap_content"
+       android:layout_height="wrap_content">
 
     <com.willowtreeapps.spurceexampleapp.widgets.RadioGroupGridLayout
         android:id="@+id/positional_radio_group"
